create headings.tsx + refact header.tsx
