+++ src/components/global/header/header.tsx
@@ -1,59 +1,8 @@
 import clsx from "clsx";
-
 import Logo from "../logo";
+import Headings from "./headings";
 import ThemeSwitcher from "./theme-switcher";
 import LocaleSwitcher from "./locale-switcher";
-
-import { useTranslations } from "next-intl";
-
-function Headings() {
-  const t = useTranslations("Global.Header.Headings");
-
-  return (
-    <header className="absolute bottom-16">
-      <h1
-        className={clsx(
-          "flex flex-col text-2xl",
-          "md:flex-row md:gap-2 md:text-3xl",
-          "lg:text-4xl"
-        )}
-      >
-        {t("headingOne")
-          .split(" ")
-          .map((str, i) => {
-            return <span key={i}>{str}</span>;
-          })}
-      </h1>
-      <h2
-        className={clsx(
-          "flex flex-col font-playfair-display text-xl text-zinc-500 dark:text-pink-500",
-          "md:flex-row md:gap-2 md:text-2xl",
-          "lg:text-3xl"
-        )}
-      >
-        {t("headingTwo")
-          .split(" ")
-          .map((str, i) => {
-            return <span key={i}>{str}</span>;
-          })}
-      </h2>
-    </header>
-  );
-}
-
-function Switchers() {
-  return (
-    <div
-      className={clsx(
-        "absolute top-16 flex flex-col items-end gap-4 self-center justify-self-end",
-        "lg:gap-8 lg:pr-4"
-      )}
-    >
-      <ThemeSwitcher />
-      <LocaleSwitcher />
-    </div>
-  );
-}
 
 export default function Header() {
   return (
@@ -68,7 +17,15 @@ export default function Header() {
         <Logo href="footer" />
         <Headings />
       </div>
-      <Switchers />
+      <div
+        className={clsx(
+          "absolute top-16 flex flex-col items-end gap-4 self-center justify-self-end",
+          "lg:gap-8 lg:pr-4"
+        )}
+      >
+        <ThemeSwitcher />
+        <LocaleSwitcher />
+      </div>
     </header>
   );
 }
